Sql_Agent_with_langchain/requirements.txt
langchain
langchain-community
langgraph
sqlalchemy>=1.4
pyodbc
python-dotenv
google-cloud-aiplatform 